- Histogram plugin's test histogram (shows tracks in an event) now uses the same widget for displaying min/max/bins as the combination histograms. - Histogram plugin does not load the combo histogram if AGeometry is not present - Possible to setup an "optional" dependency requirement for plugins. Used by histogram plugin

+++ trunk/src/ahist/AComboHistogram.h
@@ -25,7 +25,7 @@ Q_OBJECT
   
  public slots:
   void updateBins();
-  void addData(int);
+  void addData(float);
   
  private:
   QSpinBox *bins;
@@ -35,7 +35,7 @@ Q_OBJECT
   TH1F* hist;
   TQtWidget* widget;
 
-  QList<int> data;
+  QList<float> data;
 };
 
 #endif //ACOMBOHISTOGRAM_H_
